Updating speedy to work with separate sonic library.  Updating documentation.
