+++ speedy.c
