+++ speedy.h
